+++ bin/handler-mailer-ses.rb
@@ -24,14 +24,6 @@ class Mailer < Sensu::Handler
     @event['action'].eql?('resolve') ? 'RESOLVED' : 'ALERT'
   end
 
-  def prefix_subject
-    if params[:subject_prefix]
-      params[:subject_prefix] + ' '
-    else
-      ''
-    end
-  end
-
   def handle
     params = {
       mail_to: settings['mailer-ses']['mail_to'],
@@ -52,6 +44,7 @@ class Mailer < Sensu::Handler
             Status:  #{@event['check']['status']}
             Occurrences:  #{@event['occurrences']}
           BODY
+    prefix_subject = params[:subject_prefix] + ' ' if params[:subject_prefix]
     subject = "#{prefix_subject}#{action_to_string} - #{short_name}: #{@event['check']['notification']}"
 
     ses = AWS::SES::Base.new(