+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ## [Unreleased]
 ### Fixed
 - fixing timeout deprecation for ruby 2.3.0
+- fixing undefined local variable or method params error (handler-mailer-mailgun.rb/handler-mailer-ses.rb)
 
 ## [0.3.1] - 2016-05-11
 ### Added